stream: Cache pre-rendered fonts
When profiling, the slowest part is usually rendering the fonts into a
surface using `pg.Font.render_to(...)`. I figured pre-rendering the
fonts onto surfaces so that they can be quickly looked-up later would be
a net-positive. I was wrong. Performance easily dropped by ~10 FPS, give
or take.

The difference in this font caching approach was that I was rendering
chars as they were needed rather than spend a lot of time pre-rendering
all of them in advance. The dictionary structure is also shallower than
the current `HEAD` at the `refactor/prerender-fonts-with-trails` branch,
which made it more memory-efficient. (Or maybe I just didn't let it run
long enough to see the actual spike, since it climbs slower than the
alternative.

Signed-off-by: Raymond L. Rivera <ray.l.rivera@gmail.com>